+++ database/migrations/2026_03_11_015527_create_zonas_table.php
@@ -11,6 +11,7 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('zonas')) return;
         Schema::create('zonas', function (Blueprint $table) {
             $table->id();
             $table->foreignId('empresa_id')->constrained('empresas')->onDelete('cascade');
