+++ database/migrations/2026_03_11_015554_create_cobradores_table.php
@@ -11,6 +11,7 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('cobradores')) return;
         Schema::create('cobradores', function (Blueprint $table) {
             $table->id();
             $table->foreignId('empresa_id')->constrained('empresas')->onDelete('cascade');
