+++ database/migrations/2026_03_11_015617_create_metodos_pago_table.php
@@ -11,6 +11,7 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('metodos_pago')) return;
         Schema::create('metodos_pago', function (Blueprint $table) {
             $table->id();
             $table->foreignId('empresa_id')->constrained('empresas')->onDelete('cascade');
