+++ database/migrations/2026_03_11_015924_create_historial_clientes_table.php
@@ -11,6 +11,7 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('historial_clientes')) return;
         Schema::create('historial_clientes', function (Blueprint $table) {
             $table->id();
             $table->foreignId('cliente_id')->constrained('clientes')->onDelete('cascade');
